After having changed the wikipedia article on top-level domains (https://en.wikipedia.org/wiki/List_of_Internet_top-level_domains), I replaced the regular expression with an actual list of the top level domains actually defined or proposed. The trigger for all of this was the following "URL" found in a text:
"<10.ja>"
        "10.ja" URL <W:0>

This is a typo (missing space), and obvious follow-up question is how many top-level domains there are that are homonymous with sme words. Anyway, going from a regex to a list is one way of ensuring that we only get url's where there are real url's.

I did not include the internationalised country top level domains for now.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/giella-shared@149709 c7155fb1-f0a7-4240-a2fc-2600b6f42f90