Removed letter symbols from standalone symbol list
This revision removes a long list of letter symbols that were detected in the mhr (Eastern Mari) OCR process and put here. They are all in the Mari preprocess pmscript file, and should, if needed, rather be a part of a common pmscript file.

Also removed some doublet symbols.

TODO: Follow up for possible negative consequences of the change.